requirements.txt
paho-mqtt>=1.6.1
sense-hat>=2.4.0
toml>=0.10.2
